leads-update/app/requirements.txt
certifi==2022.9.24
msgpack
charset-normalizer==2.1.1
defusedxml==0.7.1
future==0.18.2
Glances==3.2.6.4
idna==3.4
kafka-python==2.0.2
kazoo==2.5.0
packaging==21.3
psutil==5.9.1
pyparsing==3.0.9
requests==2.28.1
six==1.16.0
tabulate==0.9.0
urllib3==1.26.12